api/src/main/java/com/jongho/OpenChatRoomMembershipRequest/dao/mapper/OpenChatRoomMembershipRequestMapper.java
package com.jongho.OpenChatRoomMembershipRequest.dao.mapper;

import com.jongho.OpenChatRoomMembershipRequest.domain.model.OpenChatRoomMembershipRequest;
import org.apache.ibatis.annotations.Mapper;
import org.apache.ibatis.annotations.Param;

@Mapper
public interface OpenChatRoomMembershipRequestMapper {

    public int countByRequesterIdAndStatus(@Param("requesterId") Long requesterId, @Param("status") int status);
    public boolean existsByRequesterIdAndOpenChatRoomIdAndStatus(@Param("requesterId") Long requesterId, @Param("openChatRoomId") Long openChatRoomId, @Param("status") int status);
    public void createOpenChatRoomMembershipRequest(OpenChatRoomMembershipRequest openChatRoomMembershipRequest);
}
